requirements.txt
selenium==3.141.0
PyAudio==0.2.11
matplotlib==2.0.2
opencv_python==3.4.2.17
requests==2.21.0
tweet_preprocessor==0.4.0
textblob==0.15.3
numpy==1.12.1
SpeechRecognition==3.8.1
tweepy==3.7.0
nltk==3.2.4
Pillow==6.0.0
google_api_python_client==1.7.8
preprocessor==1.1.3
scikit_learn==0.20.3
